Issue App crashing on Opening in Android Oreo 8.1 #1356
After reading through some articles, I have found out that using strict policy on certain devices will cause a DiskReadViolation that may not be related to the app's code, but due to an external library or each hardware maker's specific software.

By outting " .permitDiskReads()" we can avert this situation. In the comment there is a link to an article where this is explained in detail.